Refactor UI + search
